changed description and added some sitemaps also added /home page
