+++ public/sitemap.xml
@@ -2,22 +2,37 @@
 <urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
     <!-- Only one page -->
     <url>
-        <loc>https://www.ccenter.uz/ru</loc>
+        <loc>https://ccenter.uz/ru</loc>
         <changefreq>monthly</changefreq>
         <priority>1.0</priority>
     </url>
       <url>
-        <loc>https://www.ccenter.uz/uz</loc>
+        <loc>https://ccenter.uz/uz</loc>
         <changefreq>monthly</changefreq>
         <priority>1.0</priority>
     </url>
       <url>
-        <loc>https://www.ccenter.uz/en</loc>
+        <loc>https://ccenter.uz/en</loc>
         <changefreq>monthly</changefreq>
         <priority>1.0</priority>
     </url>
      <url>
-        <loc>https://www.ccenter.uz/</loc>
+        <loc>https://ccenter.uz/</loc>
+        <changefreq>monthly</changefreq>
+        <priority>1.0</priority>
+    </url>
+      <url>
+        <loc>https://ccenter.uz/ru/home</loc>
+        <changefreq>monthly</changefreq>
+        <priority>1.0</priority>
+    </url>
+    <url>
+        <loc>https://ccenter.uz/uz/home</loc>
+        <changefreq>monthly</changefreq>
+        <priority>1.0</priority>
+    </url>
+    <url>
+        <loc>https://ccenter.uz/en/home</loc>
         <changefreq>monthly</changefreq>
         <priority>1.0</priority>
     </url>
